[Backport 4.2.x] Move datastorage providers to maven modules and include them in the build on demand.
Backport https://github.com/geonetwork/core-geonetwork/pull/7302
