+++ datastorages/README.md
@@ -1,0 +1,12 @@
+# Data storage
+
+Extensions for data storage (used for storage resources such as record attachments and thumbnails).
+
+Core provides the default implementation for a data directory:
+
+* org.fao.geonet.resource.FileResources
+
+For more information see:
+
+* org.fao.geonet.resource.Resources
+
