+++ datastorages/cmis/README.md
@@ -1,0 +1,11 @@
+# CMIS Data storage
+
+Extension for data storage (used for storage resources such as record attachments and thumbnails).
+
+* cmis: content management interoperability services
+
+This data storage is based on the Apache Chemistry project which has been retired.
+
+Reference:
+
+* https://chemistry.apache.org/java/opencmis.html
