+++ datastorages/cmis/src/main/assembly/assembly.xml
@@ -1,0 +1,31 @@
+<assembly xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.0"
+          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+          xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.0 http://maven.apache.org/xsd/assembly-1.1.0.xsd">
+  <id>bin</id>
+  <formats>
+    <format>zip</format>
+  </formats>
+  <fileSets>
+    <fileSet>
+      <directory>${project.build.directory}/lib</directory>
+      <outputDirectory>/lib</outputDirectory>
+    </fileSet>
+  </fileSets>
+
+  <files>
+    <file>
+      <source>${project.build.directory}/${project.artifactId}-${project.version}.jar</source>
+      <outputDirectory>/lib</outputDirectory>
+    </file>
+
+    <file>
+      <source>${project.build.directory}/html/README.html</source>
+      <outputDirectory>/</outputDirectory>
+    </file>
+
+    <file>
+      <source>${project.build.directory}/html/license/LICENSE.html</source>
+      <outputDirectory>/license</outputDirectory>
+    </file>
+  </files>
+</assembly>
