+++ datastorages/jcloud/README.md
@@ -1,0 +1,10 @@
+# JCloud Data storage
+
+Extension for data storage (used for storage resources such as record attachments and thumbnails).
+
+This extension uses the Apache JCloud library for portable storage across different cloud providers.
+
+Reference:
+
+* https://jclouds.apache.org
+
