+++ datastorages/jcloud/src/main/assembly/README.md
@@ -1,0 +1,17 @@
+# JCloud data storage provider library
+
+To install the JCloud data storage provider in GeoNetwork:
+
+1. Copy the `lib` folder jar files into to `{GEONETWORK_DIR}/WEB-INF/lib` folder.
+
+2. Define the following environmental variables:
+   
+    - Use JCloud data storage provider:
+
+        ```bash
+        export GEONETWORK_STORE_TYPE=jcloud
+        ```
+     
+    - Setup JCloud connection `TODO`
+
+3. Start GeoNetwork
