+++ datastorages/s3/README.md
@@ -1,0 +1,9 @@
+# S3 Data storage
+
+Extension for data storage (used for storage resources such as record attachments and thumbnails).
+
+This extension uses Amazon Web Services Simple Storage Service (S3) data storage directly.
+
+Reference:
+
+* https://docs.aws.amazon.com/AmazonS3/latest/API/Welcome.html
