+++ datastorages/s3/src/main/java/org/fao/geonet/resources/S3Credentials.java
