+++ release/src/markdown/html/footer.html
@@ -1,0 +1,8 @@
+    </div>
+    <footer>
+      <p>
+        © Copyright 2023, Open Source Geospatial Foundation
+      </p>
+    </footer>
+  </body>
+</html>
