+++ release/src/markdown/html/header.html
@@ -1,0 +1,125 @@
+<!DOCTYPE html>
+
+<html lang="${lang}">
+<head>
+    <title>titleToken</title>
+    <meta charset="utf-8" />
+  <style>
+
+    body, html, h1, h2, h3, h4, h5, .rst-content .toctree-wrapper p.caption {
+      font-family: "Lucida Sans", "Lucida Sans Unicode", "Lucida Grande", Verdana, Arial, Helvetica, sans-serif;
+      -webkit-font-smoothing: antialiased;
+      text-rendering: optimizeLegibility;
+      color: #777;
+    }
+    /* general */
+    body {
+     font-family: 'Lucida Grande', 'Lucida Sans Unicode', 'Geneva', 'Verdana', sans-serif;
+      margin: 0;
+    }
+    h1 {
+      font-size: 32px;
+    }
+    h2 {
+      padding-bottom: .3rem;
+      border-bottom: 1px solid #eaecef;
+    }
+    h4 {
+      font-size: 110%;
+      text-decoration: underline;
+    }
+    p {
+      margin-bottom: 20px;
+      line-height: 24px;
+    }
+    a {
+      color: #003366;
+    }
+    a:hover, a:focus, a:visited {
+      color: #003366;
+      text-decoration: underline;
+    }
+    ul {
+      margin-bottom: 24px;
+    }
+    ul li {
+      margin-top: 5px;
+      margin-bottom: 5px;
+      line-height: 24px;
+    }
+    ul li p {
+      margin-top: 5px;
+      margin-bottom: 5px;
+    }
+    .clearfix:before, .clearfix:after {
+      content: " ";
+      display: table;
+    }
+    .clearfix:after {
+      clear: both;
+    }
+    /* layout */
+    header {
+      border-bottom: 1px solid #003366;
+      padding: 15px 10px 10px 10px;
+    }
+    .gn-logo {
+      float: left;
+    }
+    .gn-logo svg {
+      width: 42px;
+      height: 42px;
+    }
+    .wrapper {
+      max-width: 700px;
+      margin: 0 auto;
+    }
+    @media (max-width: 700px) {
+      .wrapper {
+        padding: 0 15px;
+      }
+    }
+    footer {
+      border-top: 1px solid #003366;
+      padding: 15px;
+      color: grey;
+    }
+    footer p {
+      margin: 0;
+    }
+  </style>
+</head>
+<body>
+  <header class="clearfix">
+    <div class="gn-logo">
+      <a href="https://geonetwork-opensource.org" target="_blank">
+
+<?xml version="1.0" encoding="UTF-8"?>
+<svg width="70px" height="70px" viewBox="0 0 70 70" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
+    <title>Logos/GeoNetwork</title>
+    <defs>
+        <radialGradient cx="52.047668%" cy="53.0601864%" fx="52.047668%" fy="53.0601864%" r="69.8119254%" gradientTransform="translate(0.5205, 0.5306), scale(0.8599, 1), translate(-0.5205, -0.5306)" id="radialGradient-1">
+            <stop stop-color="#CDCD08" offset="0%"></stop>
+            <stop stop-color="#CDCD08" offset="58.74%"></stop>
+            <stop stop-color="#998675" offset="100%"></stop>
+        </radialGradient>
+        <radialGradient cx="34.2741853%" cy="51.5249214%" fx="34.2741853%" fy="51.5249214%" r="34.5721994%" gradientTransform="translate(0.3427, 0.5152), scale(0.9825, 1), translate(-0.3427, -0.5152)" id="radialGradient-2">
+            <stop stop-color="#0D71B9" offset="0%"></stop>
+            <stop stop-color="#1C3761" offset="100%"></stop>
+        </radialGradient>
+    </defs>
+    <g id="Logos/GeoNetwork" stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
+        <g id="gn-logo" transform="translate(2, 3)" fill-rule="nonzero">
+            <g id="Group">
+                <path d="M33.582707,55.1704248 C18.5164297,55.1704248 6.10587257,52.8613512 4.47316747,49.9313725 C4.3520451,49.7152282 4.29066827,49.4956174 4.29066827,49.272948 C4.29066827,47.2536301 9.38881902,46.601119 17.1477877,46.3951702 C21.9751463,46.9956843 27.5904121,47.3072583 33.582707,47.3072583 C48.2996869,47.3072583 60.2310562,49.9340233 60.2310562,53.1739448 C60.2310562,53.4420861 60.1494923,53.6871856 59.9912584,53.9108745 C58.8991179,55.4536554 54.1657617,55.9856559 47.4920005,56.1663199 C43.7924665,55.5468422 38.9206556,55.1704248 33.582707,55.1704248 Z M17.6436961,18.6169572 C20.6215936,18.2486962 23.897811,17.8300696 27.37386,17.8300696 L27.3742678,17.8300696 C27.624465,15.3217762 27.9129973,13.6849929 28.2535265,11.855311 C29.9677955,11.9376905 31.6033554,11.9847936 33.5117464,11.9847936 L33.5117464,17.8300696 L33.58495,17.8300696 C39.3843463,17.8300696 44.8183363,18.1016773 49.5160084,18.68282 C56.0246026,17.6139253 60.2345227,15.8749832 60.2345227,13.9494638 C60.2345227,13.5901748 60.0862803,13.2229334 59.8061084,12.8830159 C57.5547412,10.1447123 46.6702444,8.06055112 33.582707,8.06055112 C20.6014066,8.06055112 9.78746252,10.1004639 7.41660424,12.8055302 C7.10034027,13.1678778 6.93374603,13.5369544 6.93374603,13.9190812 C6.93374603,15.843785 11.1422387,17.5476546 17.6436961,18.6169572 Z" id="Shape" fill="url(#radialGradient-1)"></path>
+                <path d="M59.8028458,12.8729066 C59.940907,13.0507025 60.0756009,13.2318657 60.2136621,13.413029 C60.0877233,13.2278249 59.9557233,13.0480086 59.8095805,12.8675188 L59.8028458,12.8729066 Z" id="Shape" stroke="#35679A" stroke-width="0.2674" fill="#1C3761"></path>
+                <path d="M46.6916549,64.561283 L46.6932862,64.561283 C46.4683737,63.4165338 40.5598858,63.2024286 33.582707,63.2024286 C23.7281582,63.2024286 15.4634934,61.7952476 13.2288468,60.066705 L13.2321094,60.0614033 C9.67103007,57.3210606 6.69109346,53.8727434 4.47316747,49.9362663 C6.10566866,52.8668568 18.5164297,55.1785812 33.582707,55.1785812 C38.9206556,55.1785812 43.7922626,55.5619315 47.4920005,56.1816131 C51.8097888,56.9050848 54.5313718,57.975203 54.5313718,59.149723 C54.5313718,59.4419256 54.3631463,59.7565583 54.0440276,60.0300012 L54.0436197,60.102797 C51.8032637,61.8407195 49.3351405,63.1439065 46.6916549,64.561283 Z M60.2310562,53.1737409 C60.2310562,53.4418822 60.1494923,53.6869816 59.9912584,53.9106706 C62.4410297,50.7129584 64.4680962,47.014444 65.6377224,43.055129 C65.6940014,42.8854761 65.7219371,42.6988987 65.7219371,42.5267989 C65.7219371,38.6871789 51.3330479,35.5751088 33.582707,35.5751088 C15.3464492,35.5751088 0.54810932,32.5176865 0.54810932,28.6788822 L0.545866313,28.6034356 C0.305048937,30.2226827 0.18025619,31.8800608 0.18025619,33.5659864 C0.18025619,36.0257493 0.445950552,38.4241354 0.950627102,40.7323934 C1.46733433,43.1617738 7.85664141,45.2410412 17.1479916,46.3951702 C21.9753502,46.9956843 27.590616,47.3072583 33.5829109,47.3072583 C48.2996869,47.3070544 60.2310562,49.9338194 60.2310562,53.1737409 Z M40.234038,19.6012294 C40.5992404,23.2367357 40.8265997,27.2284724 40.8863452,31.4322751 C55.1167966,30.7167559 65.7160237,27.9572457 65.7160237,24.6606373 C65.7160237,24.4881297 65.6942053,24.3141947 65.6377224,24.1455614 C64.439141,20.0616575 62.4807921,16.3036014 59.9272307,13.0328896 C60.1264505,13.3218296 60.2308523,13.6179066 60.2308523,13.9186734 C60.2310562,16.6533066 51.7339344,18.9509613 40.234038,19.6012294 Z M33.5665981,11.9847936 C31.7124471,11.9847936 29.9955272,11.9421765 28.28187,11.8604087 C27.8469305,14.1857951 27.4725523,16.7938004 27.1901373,19.6148913 C15.5642248,18.9841986 6.93863986,16.6730859 6.93863986,13.9182656 C6.93863986,13.6164792 7.04161427,13.3191788 7.2416497,13.0288114 C4.67157162,16.320322 2.67937369,20.0985652 1.48283146,24.2120359 C1.4396026,24.3606861 1.43491268,24.509948 1.43491268,24.6602295 C1.43491268,28.4992377 16.0625802,31.6115117 33.5665981,31.6115117 L33.5665981,11.9847936 Z M46.0931799,3.35064447 C46.0931799,2.75645156 45.2367591,2.2054875 43.7806398,1.75505094 L43.782271,1.74974928 C40.5664109,0.720616941 37.139912,0.163943413 33.582707,0.163943413 C29.9211002,0.163943413 26.3985598,0.752630767 23.1005241,1.84130475 L23.1021553,1.84579076 C21.8187476,2.27869109 21.072438,2.79580613 21.072438,3.35064447 C21.072438,4.87140315 26.6728184,6.10526085 33.582707,6.10526085 C40.4913721,6.10526085 46.0931799,4.87140315 46.0931799,3.35064447 Z" id="Shape" fill="url(#radialGradient-2)"></path>
+            </g>
+        </g>
+    </g>
+</svg>
+
+      </a>
+    </div>
+  </header>
+  <div class="wrapper" role="main">
